requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
Django==4.2.14
djangorestframework==3.15.2
idna==3.7
mnemonic==0.20
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyNaCl==1.5.0
requests==2.32.3
requests-sse==0.3.2
sqlparse==0.5.0
stellar-sdk==10.0.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xdrlib3==0.1.1
